requirements.txt
pandas
numpy
scipy
pyspark
pytest
scikit-learn
allure-pytest